+++ oneflow/user/kernels/dim_gather_kernel_util.cu
@@ -23,7 +23,7 @@ namespace user_op {
 template<typename IN_T, typename IDX_T>
 __global__ void DoCUDADimGather(DimOpIndexNdHelper<IDX_T> input_nd_helper,
                                 DimOpIndexNdHelper<IDX_T> index_nd_helper, int ndim,
-                                int64_t elem_cnt, int64_t dim, const IDX_T* index,
+                                int64_t elem_cnt, int32_t dim, const IDX_T* index,
                                 const IN_T* input, IN_T* output) {
   DoDimGather<IN_T, IDX_T>(input_nd_helper, index_nd_helper, ndim, elem_cnt, dim, index, input,
                            output);
@@ -33,7 +33,7 @@ template<typename IDX_T, typename IN_T>
 struct DimGatherFunctor<DeviceType::kGPU, IN_T, IDX_T> final {
   void operator()(DeviceCtx* ctx, DimOpIndexNdHelper<IDX_T> input_nd_helper,
                   DimOpIndexNdHelper<IDX_T> index_nd_helper, int ndim, int64_t elem_cnt,
-                  int64_t dim, const IDX_T* index, const IN_T* input, IN_T* output) {
+                  int32_t dim, const IDX_T* index, const IN_T* input, IN_T* output) {
     RUN_CUDA_KERNEL((DoCUDADimGather<IN_T, IDX_T>), ctx, BlocksNum4ThreadsNum(elem_cnt),
                     input_nd_helper, index_nd_helper, ndim, elem_cnt, dim, index, input, output);
   }
@@ -44,7 +44,7 @@ template<typename IDX_T>
 struct DimGatherFunctor<DeviceType::kGPU, float16, IDX_T> final {
   void operator()(DeviceCtx* ctx, DimOpIndexNdHelper<IDX_T> input_nd_helper,
                   DimOpIndexNdHelper<IDX_T> index_nd_helper, int ndim, int64_t elem_cnt,
-                  int64_t dim, const IDX_T* index, const float16* input, float16* output) {
+                  int32_t dim, const IDX_T* index, const float16* input, float16* output) {
     RUN_CUDA_KERNEL((DoCUDADimGather<half, IDX_T>), ctx, BlocksNum4ThreadsNum(elem_cnt),
                     input_nd_helper, index_nd_helper, ndim, elem_cnt, dim, index,
                     reinterpret_cast<const half*>(input), reinterpret_cast<half*>(output));
@@ -54,7 +54,7 @@ struct DimGatherFunctor<DeviceType::kGPU, float16, IDX_T> final {
 template<typename IN_T, typename IDX_T>
 __global__ void DoCUDAScatterDimAdd(DimOpIndexNdHelper<IDX_T> input_nd_helper,
                                     DimOpIndexNdHelper<IDX_T> output_nd_helper, int ndim,
-                                    int64_t elem_cnt, int64_t dim, const IDX_T* index,
+                                    int64_t elem_cnt, int32_t dim, const IDX_T* index,
                                     const IN_T* input, IN_T* output) {
   DoDimScatterAdd<IN_T, IDX_T>(input_nd_helper, output_nd_helper, ndim, elem_cnt, dim, index, input,
                                output);
@@ -64,7 +64,7 @@ template<typename IN_T, typename IDX_T>
 struct DimScatterAddFunctor<DeviceType::kGPU, IN_T, IDX_T> final {
   void operator()(DeviceCtx* ctx, DimOpIndexNdHelper<IDX_T> input_nd_helper,
                   DimOpIndexNdHelper<IDX_T> output_nd_helper, int ndim, int64_t elem_cnt,
-                  int64_t dim, const IDX_T* index, const IN_T* input, IN_T* output) {
+                  int32_t dim, const IDX_T* index, const IN_T* input, IN_T* output) {
     RUN_CUDA_KERNEL((DoCUDAScatterDimAdd<IN_T, IDX_T>), ctx, BlocksNum4ThreadsNum(elem_cnt),
                     input_nd_helper, output_nd_helper, ndim, elem_cnt, dim, index, input, output);
   }
@@ -75,7 +75,7 @@ template<typename IDX_T>
 struct DimScatterAddFunctor<DeviceType::kGPU, float16, IDX_T> final {
   void operator()(DeviceCtx* ctx, DimOpIndexNdHelper<IDX_T> input_nd_helper,
                   DimOpIndexNdHelper<IDX_T> output_nd_helper, int ndim, int64_t elem_cnt,
-                  int64_t dim, const IDX_T* index, const float16* input, float16* output) {
+                  int32_t dim, const IDX_T* index, const float16* input, float16* output) {
     RUN_CUDA_KERNEL((DoCUDAScatterDimAdd<half, IDX_T>), ctx, BlocksNum4ThreadsNum(elem_cnt),
                     input_nd_helper, output_nd_helper, ndim, elem_cnt, dim, index,
                     reinterpret_cast<const half*>(input), reinterpret_cast<half*>(output));
